requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
HackRequests==1.2
idna==2.9
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dateutil==2.8.1
requests==2.23.0
requests-toolbelt==0.9.1
selenium==3.141.0
six==1.14.0
urllib3==1.25.9
